refactor: 레이어 간 dto 의존 관계 경계 격리
