+++ src/main/java/nextstep/api/favorite/domain/service/FavoriteService.java
@@ -3,15 +3,16 @@ package nextstep.api.favorite.domain.service;
 import java.util.List;
 
 import nextstep.api.favorite.application.model.dto.FavoriteCreateRequest;
-import nextstep.api.favorite.domain.model.dto.FavoriteCreateInfo;
-import nextstep.api.favorite.domain.model.dto.FavoriteInfo;
+import nextstep.api.favorite.domain.model.dto.inport.FavoriteCreateCommand;
+import nextstep.api.favorite.domain.model.dto.outport.FavoriteCreateInfo;
+import nextstep.api.favorite.domain.model.dto.outport.FavoriteInfo;
 
 /**
  * @author : Rene Choi
  * @since : 2024/02/12
  */
 public interface FavoriteService {
-	FavoriteCreateInfo create(FavoriteCreateRequest request);
+	FavoriteCreateInfo create(FavoriteCreateCommand request);
 
 	List<FavoriteInfo> findFavorites(Long memberId);
